requirements.txt
alembic==1.8.1
click==8.1.3
Flask==2.2.2
Flask-Migrate==4.0.0
Flask-SQLAlchemy==3.0.2
importlib-metadata==5.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
load-dotenv==0.1.0
Mako==1.2.4
MarkupSafe==2.1.1
python-dotenv==0.21.0
SQLAlchemy==1.4.45
Werkzeug==2.2.2
zipp==3.11.0
